pyproject.toml
[project]
name = "link-market"
version = "0.1.0"
description = "Hyperliquid LINK perpetual market monitoring system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]
